refact: fix formatting and types

+++ apps/mobile/app/(authenticated)/(tabs)/(signer,explorer,converter)/account/[id]/signAndSend/previewMessage.tsx
@@ -1458,7 +1458,7 @@ function PreviewMessage() {
       }
 
       if (result.txData) {
-        const txHex = Array.from(result.txData)
+        const txHex = Array.from(result.txData as Uint8Array)
           .map((b) => b.toString(16).padStart(2, '0'))
           .join('')
